+++ packages/blob/src/debug.ts
@@ -8,14 +8,13 @@ try {
   ) {
     debugIsActive = true;
   }
-} catch (error) {
+} catch {
   // noop
 }
 
 // Set process.env.DEBUG = 'blob' to enable debug logging
 export function debug(message: string, ...args: unknown[]): void {
   if (debugIsActive) {
-    // eslint-disable-next-line no-console -- Ok for debugging
     console.debug(`vercel-blob: ${message}`, ...args);
   }
 }