+++ packages/edge-config/tsup.config.js
@@ -1,14 +1,27 @@
 import { defineConfig } from 'tsup';
 
-// eslint-disable-next-line import/no-default-export -- [@vercel/style-guide@5 migration]
-export default defineConfig({
-  entry: ['src/index.ts'],
-  format: ['esm', 'cjs'],
-  splitting: true,
-  sourcemap: true,
-  minify: false,
-  clean: true,
-  skipNodeModulesBundle: true,
-  dts: true,
-  external: ['node_modules'],
-});
+export default [
+  defineConfig({
+    entry: ['src/index.ts'],
+    format: ['esm', 'cjs'],
+    splitting: true,
+    sourcemap: true,
+    minify: false,
+    clean: true,
+    skipNodeModulesBundle: true,
+    dts: true,
+    external: ['node_modules'],
+  }),
+  // Separate configs so we don't get split types
+  defineConfig({
+    entry: ['src/index.next-js.ts'],
+    format: ['esm', 'cjs'],
+    splitting: true,
+    sourcemap: true,
+    minify: false,
+    clean: true,
+    skipNodeModulesBundle: true,
+    dts: true,
+    external: ['node_modules'],
+  }),
+];