+++ test/next/src/app/vercel/blob/handle-blob-upload.ts
@@ -1,4 +1,4 @@
-import { handleUpload, type HandleUploadBody } from '@vercel/blob/client';
+import { type HandleUploadBody, handleUpload } from '@vercel/blob/client';
 import { NextResponse } from 'next/server';
 import { validateUploadToken } from './validate-upload-token';
 
@@ -58,15 +58,6 @@ export async function handleUploadHandler(
             customHeader,
           }),
         };
-      },
-      // eslint-disable-next-line @typescript-eslint/require-await -- [@vercel/style-guide@5 migration]
-      onUploadCompleted: async ({ blob, tokenPayload }) => {
-        console.log('Upload completed', blob, tokenPayload);
-        try {
-          //   await db.update({ avatar: blob.url, userId: tokenPayload.userId });
-        } catch (error) {
-          throw new Error('Could not update user');
-        }
       },
     });